+++ src/components/Cast/index.js
@@ -1,3 +1,4 @@
+import Content from "../../styledComponents/Content/index.js";
 import { TilesPersonSection } from "./styled";
 import TilePerson from "./tilePerson.js";
 
@@ -44,6 +45,7 @@ const Cast = ({ selectedCast }) => {
         with desktop publishing software like Aldus PageMaker including versions
         of Lorem Ipsum.
       </p>
+      <Content/>
       <TilesPersonSection>
         {selectedCast.map((person) => (
           <TilePerson
