+++ src/styledComponents/StyledLogos/styled.js
